requirements.txt
python-dotenv
mistralai
streamlit
pybase64